CONF-1 Add webrtc adapter. Finish implementation of changing audio/video input
