+++ package.json
@@ -11,7 +11,8 @@
   },
   "dependencies": {
     "loglevel": "^1.6.8",
-    "uuid": "^8.0.0"
+    "uuid": "^8.0.0",
+    "webrtc-adapter": "^7.6.1"
   },
   "peerDependencies": {},
   "devDependencies": {
